Mobilify

+++ app/src/webgl/common.ts
@@ -26,7 +26,7 @@ export type TextureInfo = {
     [key: string]: TextureObject;
 };
 
-export type UniformSetter<T> = number | Float32Array | Int32Array | ((shape: T) => number | Float32Array | Int32Array)
+export type UniformSetter<T> = number | Float32Array | Int16Array | ((shape: T) => number | Float32Array | Int16Array)
 export type UniformObject<T> = Record<string, UniformSetter<T>>;
 
 export interface DrawableScene<T> {
